Rename `node-input.tsx` to `node-input-components.tsx`
